update user model meta, move user manager to models.py
